added routing to popular page
