+++ frontend/package.json
@@ -2,7 +2,7 @@
   "name": "frontend",
   "version": "1.0.0",
   "private": true,
-  "main": "index.ts",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -20,10 +20,16 @@
     "@expo-google-fonts/instrument-sans": "^0.4.2",
     "@expo-google-fonts/league-spartan": "^0.4.2",
     "expo": "~54.0.30",
+    "expo-constants": "~18.0.13",
     "expo-font": "~14.0.10",
+    "expo-linking": "~8.0.11",
+    "expo-router": "~6.0.23",
     "expo-status-bar": "~3.0.9",
     "react": "19.1.0",
+    "react-dom": "19.1.0",
     "react-native": "0.81.5",
+    "react-native-safe-area-context": "~5.6.0",
+    "react-native-screens": "~4.16.0",
     "react-native-svg": "15.12.1"
   },
   "devDependencies": {
