+++ frontend/src/app/popular.tsx
@@ -1,0 +1,17 @@
+import { View } from "react-native";
+
+import { ActivityList } from "../components/ActivityList";
+import { mockActivities } from "../data/mockActivities";
+import { HomePopUpPageHeaderSection } from "../home_components/HomePopUpPageHeaderSection";
+
+export default function PopularScreen() {
+  return (
+    <View>
+      <HomePopUpPageHeaderSection
+        sectionName="Popular"
+        rightPadding={202}
+      ></HomePopUpPageHeaderSection>
+      <ActivityList header="" activities={mockActivities} variant="card" horizontal={false} />
+    </View>
+  );
+}
